[dagit] Export the new Workspace grid for Cloud use
